% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corpus_ITA}
\alias{corpus_ITA}
\title{Example Corpus: Italian Parliamentary Speeches (2018–2022)}
\format{
A \code{quanteda::tokens} object with document-level variables
including \code{Speaker_party} and \code{year_month}.
}
\source{
https://clarin.si/repository/xmlui/handle/11356/1864
}
\usage{
data(corpus_ITA)
}
\description{
A tokenized \code{quanteda} tokens object from Italian parties.
}
\keyword{datasets}
