% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_cossim_IT}
\alias{demo_cossim_IT}
\title{Demo Cosine Similarity Results (Italy)}
\format{
A list of length 1. The list contains a data frame with the
}
\source{
Generated using \code{calc_cosSim()} with Italian parliamentary text data
}
\usage{
data(demo_cossim_IT)
}
\description{
Example cosine similarity results generated by \code{calc_cosSim()}
for Italian political party word embeddings (FdI, 2018),
focusing on immigration-related terms.
}
\keyword{datasets}
