% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_visualisation.R
\name{plot_libVillib_wordfacet}
\alias{plot_libVillib_wordfacet}
\title{Plot Similarity Scores for Words in a Political Context}
\usage{
plot_libVillib_wordfacet(country = "Italy", model = "model1")
}
\description{
Plot Similarity Scores for Words in a Political Context
}
