% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embeddings.R
\name{train_word_embeddings}
\alias{train_word_embeddings}
\title{Train word embeddings from a corpus}
\usage{
train_word_embeddings(
  corpus = NULL,
  model_name = "model1",
  country = "Italy",
  speaker_party = "Speaker_party",
  num_bootstraps = 100,
  window_size = 10,
  drop_parties = c("RE"),
  min_docs = 50,
  output_dir = "./models_wordembeddings/"
)
}
\arguments{
\item{corpus}{A quanteda tokens object. If NULL, uses internal data for Italy.}

\item{model_name}{A custom name to label your saved models.}

\item{country}{A string indicating the country (used in folder structure).}

\item{speaker_party}{The docvar indicating actor name.}

\item{num_bootstraps}{Number of bootstrap samples per year.}

\item{window_size}{The size of the context window.}

\item{drop_parties}{A vector of actors to exclude.}

\item{min_docs}{Minimum number of documents required per year.}

\item{output_dir}{Path to the directory for storing models (relative to working directory).}
}
\value{
Saves RDS files with word embeddings; returns nothing.
}
\description{
This function trains GloVe-based word embeddings for individual speakers across years,
using a bootstrapping procedure on a quanteda corpus object.
}
\examples{
\dontrun{
train_word_embeddings(corpus = your_tokens_object)
}
}
