% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpus.R
\docType{data}
\name{corpus_ITA}
\alias{corpus_ITA}
\title{Example Corpus: Italian Parliamentary Speeches (2018–2022)}
\format{
A \code{quanteda} tokens object with docvars: \code{Speaker_party}, \code{year_month}, \code{Speaker_party}.
}
\usage{
data(corpus_HU_8parties_2014_2023)
}
\description{
A tokenized \code{quanteda} tokens object from Italian parties.
}
\keyword{datasets}
